// src/app/model/equipement.ts
export class Equipement {
  id?: number;
  serialNumber?: String;
  localisation?: string;
  host?: string;
  ipAddress?: string;
  image?: string;
  model?: string;
  actualOS?: string;
  recolOS?: string;
  eos?: Date;
  eol?: Date;
  constructer?: string;
  productGroup?: string;
  createdBy?: any;
  UserInformations?: string;
  updatedBy?: any[];
  contratId?: number;
  lastUpdate?: any;
}